+++ src/Form/DashboardActionType.php
@@ -14,6 +14,8 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
  *
  * @author Héctor Franco Aceituno <hectorfranco@nowo.tech>
  * @copyright 2026 Nowo.tech
+ *
+ * @extends AbstractType<array<string, mixed>>
  */
 #[FormKitConfig('dashboard_menu')]
 final class DashboardActionType extends AbstractType