+++ src/Form/MenuDefinitionType.php
@@ -20,6 +20,8 @@ use Symfony\Contracts\Translation\TranslatorInterface;
  *
  * @author Héctor Franco Aceituno <hectorfranco@nowo.tech>
  * @copyright 2026 Nowo.tech
+ *
+ * @extends AbstractType<Menu>
  */
 #[FormKitConfig('dashboard_menu')]
 final class MenuDefinitionType extends AbstractType